Refactor Fireblocks CW service structure
