+++ src/providers/fireblocks/cw/core/services/cw-deposit.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { FireblocksClientService } from '../shared/fireblocks-client.service';
+import { FireblocksCwService } from '../../fireblocks-cw.service';
 
 @Injectable()
 export class CwDepositService {
   private readonly logger = new Logger(CwDepositService.name);
 
-  constructor(private readonly client: FireblocksClientService) {}
+  constructor(private readonly client: FireblocksCwService) {}
 
   async activateAssetInVault(vaultAccountId: string, assetId: string): Promise<void> {
     this.logger.log(`Activate asset ${assetId} in vault ${vaultAccountId}`);
