+++ src/providers/fireblocks/cw/core/services/cw-portfolio.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { FireblocksClientService } from '../shared/fireblocks-client.service';
+import { FireblocksCwService } from '../../fireblocks-cw.service';
 
 @Injectable()
 export class CwPortfolioService {
   private readonly logger = new Logger(CwPortfolioService.name);
 
-  constructor(private readonly client: FireblocksClientService) {}
+  constructor(private readonly client: FireblocksCwService) {}
 
   async getBalances(vaultAccountId: string): Promise<void> {
     this.logger.log(`Fetch balances for vault ${vaultAccountId}`);
