+++ src/providers/fireblocks/cw/core/services/cw-transfers.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { FireblocksClientService } from '../shared/fireblocks-client.service';
+import { FireblocksCwService } from '../../fireblocks-cw.service';
 import { FireblocksErrorMapper } from '../../infrastructure/persistence/relational/mappers/fireblocks-error.mapper';
 import { FireblocksResilienceService } from '../shared/fireblocks-resilience.service';
 
@@ -17,7 +17,7 @@ export class CwTransfersService {
   private readonly logger = new Logger(CwTransfersService.name);
 
   constructor(
-    private readonly client: FireblocksClientService,
+    private readonly client: FireblocksCwService,
     private readonly errorMapper: FireblocksErrorMapper,
     private readonly resilience: FireblocksResilienceService,
   ) {}
